+++ packages/network-controller/src/rpc-service/rpc-service-chain.ts
@@ -21,6 +21,7 @@ import type {
 import { projectLogger, createModuleLogger } from '../logger';
 
 const log = createModuleLogger(projectLogger, 'RpcServiceChain');
+// const log = console.log.bind(console);
 
 /**
  * Statuses that the RPC service chain can be in.