+++ packages/network-controller/src/rpc-service/rpc-service.ts
@@ -61,6 +61,7 @@ export type RpcServiceOptions = {
 };
 
 const log = createModuleLogger(projectLogger, 'RpcService');
+// const log = console.log.bind(console);
 
 /**
  * The maximum number of times that a failing service should be re-run before